apps/agent-api/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
distro==1.9.0
executing==2.1.0
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpickle==1.4.2
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langgraph==0.2.62
langgraph-checkpoint==2.0.9
langgraph-checkpoint-postgres==2.0.10
langgraph-sdk==0.1.51
langsmith==0.2.10
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.1
openai==1.59.6
orjson==3.10.14
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
redisbeat==1.2.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rq==2.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
